backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
openai==1.59.7
google-genai==0.6.0
sentence-transformers==3.3.1
umap-learn==0.5.7
numpy==2.2.1
scikit-learn==1.6.0
llama-index==0.12.10
llama-index-vector-stores-chroma==0.4.1
chromadb==0.5.23
musicbrainzngs==0.7.1
requests==2.32.3
